libgambatte: compilation and warning fixes for clang.
git-svn-id: https://gambatte.svn.sourceforge.net/svnroot/gambatte@276 9dfb2916-2d38-0410-aef4-c5fe6c9ffc24

+++ libgambatte/src/gambatte.cpp
@@ -64,7 +64,7 @@ long GB::runFor(gambatte::uint_least32_t *const videoBuf, const int pitch,
 	const long cyclesSinceBlit = p_->cpu.runFor(samples * 2);
 	samples = p_->cpu.fillSoundBuffer();
 	
-	return cyclesSinceBlit < 0 ? cyclesSinceBlit : samples - (cyclesSinceBlit >> 1);
+	return cyclesSinceBlit < 0 ? cyclesSinceBlit : static_cast<long>(samples) - (cyclesSinceBlit >> 1);
 }
 
 void GB::reset() {
